pull up ConvertStringToBytesref
